FSM multiplication simplified
It never quite made sense for multiplication of a finite state machine
to require a lower and upper bound, with the upper bound possibly being
infinite. The properties of the possibly-unlimited "multiplier" object
are specific to the lego module and shouldn't expand elsewhere.
fsm.__mul__() now accepts only a non-negative integer as the multiplier,
not a bizarre min/max tuple, whereas lego.mult.fsm() handles the task of
multiplying an FSM by a "multiplier" object.