lego.fsm() alphabet argument now optional
There is no real reason to force the user to supply an alphabet when
calling lego.fsm().

It would be nice if there was a cool Pythonic way to set this default
argument to self.alphabet() each time, like "def fsm(self,
alphabet=self.alphabet())" (which doesn't work) or by overloading the
lego.fsm() method using "def fsm(self): return fsm(self,
self.alphabet())", but neither of these worked. Maybe there is another
way to do this that I'm not aware of.