"|.+" now reduces() to ".*"
I discovered that lego.parse("|.+") has a fairly obvious reduction path
which wasn't implemented. It will now reduce to ".*".

As a consequence, something like "|a+|b+" will now reduce() to either
"a*|b+" or to "a+|b*". It will not reduce() to "a*|b*", because I'm not
certain that this is desirable. If we carry on down this path, we can
end up reducing() "a*|b{1, 5}" to "a*|b{0,5}" and it's debatable which
of these is the reduced() version of the other.