Discard charclass subtraction
I guess we never actually needed to be able to subtract one charclass
from another, as amusing an academic exercise as it might be. In any
case, this usage of subtraction is inconsistent with the rest of the
module where subtraction is considered to be the opposite of
concatenation (i.e. addition).