import changes
* Took "import *" out of __init__py. This is risky in principle, and I
also feel that fsm.py and lego.py should remain as separate packages.
Although they are coupled together, this is only the case if you want to
convert one type of object to the other. They can be imported and used
separately.
* Remove modifications to sys.path from example code and elsewhere.
* Example code is now stored in the top level directory whereas
greeenery itself has a directory of its own. To make greenery work, put
it in with the rest of your python packages, or add greenery's parent
directory to your PYTHONPATH.
* Removing absolute_import import calls - all this code works fine
without it.