Move licence to somewhere more prominent, +README
Someone raised an issue because they weren't able to find the licence
for this project, so I've moved it out of the code (where, honestly, it
was just an eyesore taking up space) and into a dedicated file at the
top level where everybody can see it.

I've also added a significant quantity of API documentation to README
and renamed it to an MD file, because just a trio of hyperlinks doesn't
really cut it on GitHub anymore.