Remove asserts from live code
We now raise honest `Exception` instances if invalid values are supplied
to constructors or methods. It's a little more verbose, and better
exception class choices than might be a future improvement. Also a few
tests.

Additionally, a new `reduce()` method is added, "a|a*" to "a*", with
some rather more complex possibilities besides. Plus unit tests. As part
of this, `multiplier`s have a new `__or__()` method and a corresponding
`canunion()` test.