Accept "any other character"
If the alphabet for an FSM contains `lego.otherchars`, e.g. `{"a", "b",
"c", lego.otherchars}` and you pass in an unknown symbol e.g. `"d"`, it
is treated as `lego.otherchars`.

It is becoming increasingly clear that the constant `lego.otherchars`
should have been placed in the `fsm` module, not the `lego` module. This
may change soon.