fix issue #43: improve performance of from_fsm
The idea is to not wait till the very end with reducing, because the way
that the equations are solved leads to very complex *repeated* lego
pieces, that then have to be reduced in many many different places.
Instead, repeatedly reduce after the substitutions to prevent the
complexity from piling up.