pylint: suppress 'arguments-differ' on python 3.8 and 3.9
The type declaration of Generic.__new__ is over-broad in Python 3.8
(likely also in 3.9, though not verified), but fixed in later versions.

This causes warnings if `Generic` subclasses override `__new__` but
don't expose an "anything goes" (*args, **kwargs) signature.

It's not worth writing a specialized transformer to fix it, and it's
lame to have to write lint suppressions on correct code, especially when
it only applies to a particular version.