lint: remove some useless suppressions
And enables the `useless-suppression` checker, which is disabled by
default.

Removes some statements which would not trigger warnings, for some
reason or other: generally these are left over from refactors.

For example, the too-many-arguments on `Fsm.__init__` had been
triggering before the arguments were explicitly marked as keyword-only,
but the suppression was not removed with that change.

pylint: useless-suppression